+++ src/pages/User/LogIn.vue
@@ -4,8 +4,7 @@
       class="customToolbar"
       :pageTitle="$t('lang.pages.login.toolbar.title')"
       :backLabel="volver"
-    >
-    </the-custom-toolbar>
+    ></the-custom-toolbar>
     <div class="container">
       <!-------------------- LOG IN FORM ------------------------------------------------>
 
@@ -13,16 +12,9 @@
         <v-ons-list class="form-list">
           <!-- EMAIL INPUT -->
 
-          <v-ons-list-item
-            :modifier="md ? 'nodivider' : ''"
-            class="form__input"
-          >
+          <v-ons-list-item :modifier="md ? 'nodivider' : ''" class="form__input">
             <div class="left">
-              <v-ons-icon
-                icon="ion-ios-email, material:zmdi-email"
-                class="list-item__icon"
-              >
-              </v-ons-icon>
+              <v-ons-icon icon="ion-ios-email, material:zmdi-email" class="list-item__icon"></v-ons-icon>
             </div>
             <div class="center">
               <v-ons-input
@@ -33,23 +25,15 @@
                 float
                 modifier="underbar"
                 v-model="email"
-              >
-              </v-ons-input>
+              ></v-ons-input>
             </div>
           </v-ons-list-item>
 
           <!-- PASSWORD INPUT -->
 
-          <v-ons-list-item
-            :modifier="md ? 'nodivider' : ''"
-            class="form__input"
-          >
+          <v-ons-list-item :modifier="md ? 'nodivider' : ''" class="form__input">
             <div class="left">
-              <v-ons-icon
-                icon="ion-locked, material:zmdi-lock"
-                class="list-item__icon"
-              >
-              </v-ons-icon>
+              <v-ons-icon icon="ion-locked, material:zmdi-lock" class="list-item__icon"></v-ons-icon>
             </div>
             <div class="center">
               <v-ons-input
@@ -63,23 +47,20 @@
                 v-model="password"
                 v-bind="$attrs"
                 ref="passwordInput"
-              >
-              </v-ons-input>
+              ></v-ons-input>
               <div class="right">
                 <v-ons-icon
                   v-if="!passwordVisible"
                   icon="ion-eye-disabled, material:zmdi-eye-off"
                   class="list-item__icon"
                   @click.prevent="togglePassword"
-                >
-                </v-ons-icon>
+                ></v-ons-icon>
                 <v-ons-icon
                   v-if="passwordVisible"
                   icon="ion-eye, material:zmdi-eye"
                   class="list-item__icon"
                   @click.prevent="togglePassword"
-                >
-                </v-ons-icon>
+                ></v-ons-icon>
               </div>
             </div>
           </v-ons-list-item>
@@ -93,9 +74,10 @@
           <!-- FORGOT PASSWORD -->
 
           <v-ons-list-item class="text__button">
-            <p class="forgotPassword" @click.prevent="onForgotPassword">
-              {{ $t('lang.pages.login.main.text1') }}
-            </p>
+            <p
+              class="forgotPassword"
+              @click.prevent="onForgotPassword"
+            >{{ $t('lang.pages.login.main.text1') }}</p>
           </v-ons-list-item>
         </v-ons-list>
       </form>
@@ -110,16 +92,12 @@
         :disabled="buttonActive"
         ripple="true"
         @click.prevent="onLognIn"
-      >
-        {{ $t('lang.pages.login.button') }}
-      </v-ons-button>
+      >{{ $t('lang.pages.login.button') }}</v-ons-button>
 
       <!------ LOGIN WITH SOCIAL BUTTONS ------>
 
       <div class="socialText">
-        <p class="socialButtons__text">
-          {{ $t('lang.pages.login.main.socialText') }}
-        </p>
+        <p class="socialButtons__text">{{ $t('lang.pages.login.main.socialText') }}</p>
         <v-ons-row class="socialButtons__list">
           <!--sign-up-button
 	          class="socialButtonsList__item-button"
@@ -144,9 +122,7 @@
 
       <!-- SIGNUP BUTTON -->
       <div>
-        <p class="text__button" @click.prevent="toSignUp">
-          {{ $t('lang.pages.login.main.text2') }}
-        </p>
+        <p class="text__button" @click.prevent="toSignUp">{{ $t('lang.pages.login.main.text2') }}</p>
       </div>
     </div>
   </v-ons-page>
@@ -171,7 +147,7 @@ export default {
   },
   created() {
     console.log('Estoy en LogIn.created')
-    // this.$store.commit('navigator/pop')
+    // this.$store.commit('navigator/POP')
   },
   mounted() {
     console.log('Estoy en LogIn.mounted')
@@ -214,7 +190,7 @@ export default {
       console.log('He olvidado la contraseña')
     },
     toSignUp() {
-      this.$store.commit('navigator/replace', SignUp)
+      this.$store.commit('navigator/REPLACE', SignUp)
     },
     socialLogIn(index) {
       this.$store.dispatch('social/dispatchLogUp', index)