+++ src/store/modules/navigator/getters.js
@@ -1,0 +1,11 @@
+export const pageStack = state => {
+  return state.stack
+}
+
+export const options = state => {
+  return state.options
+}
+
+export const index = state => {
+  return state.index
+}